draw: Correct grid draw.
The drawing primitives needs to happen at the center of the cell rather
than at the exact cell coordinates. Otherwise we get for example circles
centered on the cell's corner rather than it's center. This means all grid
coordinates need to be translated to screen coordinates which are in the
center of the cell.